+++ src/main.c
@@ -1,81 +1,4 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <time.h>
-#include "crossplatform.h"
-#include <sys/time.h>
-#include "framebuffer.h"
-#include "menu.h"
-#include <math.h>
-
-int map(int x, int inMin, int inMax, int outMin, int outMax);
-void kb_control();
-void ray_cast();
-void draw_3d();
-void draw_map();
-void draw_menu();
-
-#define WORLDSIZE 20
-struct level {
-    int world[WORLDSIZE][WORLDSIZE]; // birds perspective of world (world needs a boarder!)
-    float spX;  // spawn coordinates...
-    float spY;
-    int spA;  // ...and rotation 
-};
-struct level lvl = { // current level lvl
-    {
-        {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 1, 1, 1, 1, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 1, 1, 1, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 1, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 0, 0, 0, 0, 0, 0, 1, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 1},
-        {1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1},
-    },
-    10,
-    10,
-    180
-};
-
-struct buffer fb;
-
-// menu
-unsigned short isMenu = 1;
-int loop = 1; // loop until exit
-
-// setup array for distance to wall for each vertical display line
-unsigned short *distance;
-
-// variables for time stats
-struct timeval tNow, tLast;
-unsigned long tTaken, frame = 0;
-
-int key; // buffer for pressed key
-float pX; // player position X
-float pY; // player position Y
-int pA; // player rotation in degree (180 = north)
-float pNX, pNY, pRad; // new theoretical player position x y and rotation in radians
-
-unsigned short mapS = 10; // minimap size // must be smaller then WORLDSIZE
-const short fov = 90; //set field of view to 90 degree
-const short maxVDist = 20; //set max viewing distran
-const float stepXY = 0.5; // player step size
-const int stepA = 5;  // player rotation steps
-int cliX = 144; // default fallback windows size 
-int cliY = 48;
-double cliA;
+#include "quest.h"
 
 int main(int argc, char const *argv[]) { 
     // output character dictionary
